fix type for us and timeout when no interrupt is triggered

+++ idDHTLib.h
@@ -79,7 +79,7 @@ class idDHTLib
     volatile byte bits[5];
     volatile byte cnt;
     volatile byte idx;
-    volatile int us;
+    volatile unsigned long us;
     int intNumber;
     int pin;
     volatile float hum;
